+++ src/main/java/com/wasd/web/service/UserService.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface UserService {
     List<UserResponse> findAll();
-    UserResponse findById(Integer id);
+    UserResponse findById(Long id);
     UserResponse create(UserRequest userRequest);
-    UserResponse update(Integer id, UserRequest userRequest);
-    void delete(Integer id);
+    UserResponse update(Long id, UserRequest userRequest);
+    void delete(Long id);
 }
